moved dependency injection to utils
